Added user input for the second menu
The second menu is completed and now the third and final menu is going to take the values of the knowns. For example if you checked velocity then it will ask you beginning velocity, ending velocity, and change in velocity. You can answer with nothing if it's an unknown and with the knowns it will spit out what you are looking for.